+++ arducate/src/App.js
@@ -1,6 +1,6 @@
 // src/App.js
 import React from 'react';
-import ARCanvas from './components/ARCanvas';
+import Canvas from './components/Canvas';
 import ARControls from './components/ARControls';
 import Toolbar from './components/Toolbar';
 
@@ -8,7 +8,7 @@ const App = () => {
   return (
     <div className="flex flex-col items-center p-4">
       <Toolbar />
-      <ARCanvas />
+      <Canvas />
       <ARControls />
     </div>
   );
